[FIX] fixed unix renderer

+++ CaveEngine/Graphics/Public/Sprite/Vertex.h
@@ -111,15 +111,15 @@ namespace cave
 	{
 	}
 
-	constexpr uint8_t VERTEX_POSITION = 0b01;
-	constexpr uint8_t VERTEX_TEXCOORD = 0b10;
+	constexpr uint32_t VERTEX_POSITION = 0b01;
+	constexpr uint32_t VERTEX_TEXCOORD = 0b10;
 
 
 	typedef struct Vertex
 	{
 	public:
 		Float3 Position = Float3(0.0f, 0.0f, 0.0f);
-		uint8_t VertexFlag = VERTEX_POSITION;
+		uint32_t VertexFlag = VERTEX_POSITION;
 
 		Vertex() = default;
 		Vertex(const Vertex& other) = default;
